Add a 'typefile' command
This enables a commandline like:

vncdo -s server -p password typefile abc.txt

to type the contents of abc.txt into the target server

Limitations:

This is only intended to support typeable characters, so trying to type a binary file
will fail.

This required adding of mappings for \n and \t to the relevant keys, which was enough
for my requirements, but there may be others that need to be added for full coverage.